+++ geocaching-api/src/main/java/com/arcao/geocaching4locus/data/api/model/request/query/filter/RadiusFilter.kt
@@ -1,6 +1,6 @@
 package com.arcao.geocaching4locus.data.api.model.request.query.filter
 
-class RadiusFilter(private val distance: Float, private val unit : DistanceUnit) : Filter {
+class RadiusFilter(private val distance: Int, private val unit : DistanceUnit) : Filter {
     override fun isValid() = distance > 0
 
     override fun toString(): String {
